Manager reads command line arguments to configure Server.
There are no longer any defaults when starting the server, all arguments
must be passed from the command line.

The path must include a forward-slash at the end. TODO: check for the
forward slash when reading the path.

The file name must include a file extension if needed.

TODO: check if the port is in a valid range before continuing.