Server now communicates in LocationObject.
Server will capture a LocationObject; if no key was found, Server will
insert a key in the form of "$TIME | $IP". The LocationObject is then
inserted into the DB, and passed back to the client.

TODO: when inserting, remove old values with the same key.
TODO: return the closest location with a different key (within bounds)